+++ apps/frontend/src/libs/modules/store/libs/types/root-reducer.type.ts
@@ -1,5 +1,7 @@
 import { type reducer as appReducer } from "~/modules/app/app.js";
 import { type reducer as authReducer } from "~/modules/auth/auth.js";
+import { type reducer as exploreReducer } from "~/modules/explore/explore.js";
+import { type reducer as locationReducer } from "~/modules/location/location.js";
 import {
 	type pointOfInterestDetailsReducer,
 	type pointsOfInterestReducer,
@@ -9,6 +11,8 @@ import { type reducer as routeReducer } from "~/modules/routes/routes.js";
 type RootReducer = {
 	app: ReturnType<typeof appReducer>;
 	auth: ReturnType<typeof authReducer>;
+	explore: ReturnType<typeof exploreReducer>;
+	location: ReturnType<typeof locationReducer>;
 	pointOfInterestDetails: ReturnType<typeof pointOfInterestDetailsReducer>;
 	pointsOfInterest: ReturnType<typeof pointsOfInterestReducer>;
 	route: ReturnType<typeof routeReducer>;