+++ apps/frontend/src/modules/routes/libs/types/types.ts
@@ -1,5 +1,6 @@
 export {
 	type RouteConstructRequestDto,
 	type RouteConstructResponseDto,
+	type RouteFindAllOptions,
 	type RouteGetByIdResponseDto,
 } from "@smartscapes/shared";