+++ apps/frontend/src/modules/routes/slices/actions.ts
@@ -5,6 +5,7 @@ import { type APIResponse, type AsyncThunkConfig } from "~/libs/types/types.js";
 import {
 	type RouteConstructRequestDto,
 	type RouteConstructResponseDto,
+	type RouteFindAllOptions,
 	type RouteGetByIdResponseDto,
 } from "../libs/types/types.js";
 import { name as sliceName } from "./route.slice.js";
@@ -33,4 +34,14 @@ const getRouteById = createAsyncThunk<
 	return await routeApi.getRouteById(id);
 });
 
-export { construct, getRouteById };
+const getAll = createAsyncThunk<
+	APIResponse<RouteGetByIdResponseDto[]>,
+	RouteFindAllOptions | undefined,
+	AsyncThunkConfig
+>(`${sliceName}/get-all`, async (options, { extra }) => {
+	const { routeApi } = extra;
+
+	return await routeApi.getAll(options);
+});
+
+export { construct, getAll, getRouteById };