+++ apps/frontend/src/modules/routes/slices/route.slice.ts
@@ -7,7 +7,7 @@ import {
 	type RouteConstructResponseDto,
 	type RouteGetByIdResponseDto,
 } from "../libs/types/types.js";
-import { construct, getRouteById } from "./actions.js";
+import { construct, getAll, getRouteById } from "./actions.js";
 
 type State = {
 	dataStatus: ValueOf<typeof DataStatus>;
@@ -41,6 +41,16 @@ const { actions, name, reducer } = createSlice({
 			state.dataStatus = DataStatus.FULFILLED;
 		});
 		builder.addCase(construct.rejected, (state) => {
+			state.dataStatus = DataStatus.REJECTED;
+		});
+
+		builder.addCase(getAll.pending, (state) => {
+			state.dataStatus = DataStatus.PENDING;
+		});
+		builder.addCase(getAll.fulfilled, (state) => {
+			state.dataStatus = DataStatus.FULFILLED;
+		});
+		builder.addCase(getAll.rejected, (state) => {
 			state.dataStatus = DataStatus.REJECTED;
 		});
 	},