+++ packages/shared/src/modules/routes/libs/types/route-find-all-options.type.ts
@@ -1,5 +1,7 @@
 type RouteFindAllOptions = {
 	categories?: string | string[];
+	latitude?: number;
+	longitude?: number;
 	name?: string;
 };